move plotting to new file
