+++ src/kete/__init__.py
@@ -14,6 +14,7 @@ from . import (
     population,
     flux,
     mpc,
+    plot,
     irsa,
     ztf,
     fov,
@@ -71,6 +72,7 @@ __all__ = [
     "wise",
     "neos",
     "mpc",
+    "plot",
     "spice",
     "SimultaneousStates",
     "propagate_n_body",
